[WIP] Timers

+++ src/haven/TimerEditWnd.java
@@ -59,10 +59,9 @@ public class TimerEditWnd extends Window {
             add = new Button(60, "Add") {
                 @Override
                 public void click() {
-                    int hours = Integer.parseInt(txthours.text == "" ? "0" : txthours.text);
-                    int minutes = Integer.parseInt(txtminutes.text == "" ? "0" : txtminutes.text);
+                    long hours = Long.parseLong(txthours.text == "" ? "0" : txthours.text);
+                    long minutes = Long.parseLong(txtminutes.text == "" ? "0" : txtminutes.text);
                     long duration = (60 * hours + minutes) * 60 * 1000;
-
                     int y = 0;
                     List<TimerWdg> timers = Glob.timersThread.getall();
                     for (TimerWdg timer : timers) {
